% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockr.ui-dependencies.R
\name{add_blockr.ui_deps}
\alias{add_blockr.ui_deps}
\title{blockr.ui dependencies utils}
\usage{
add_blockr.ui_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches blockr.ui dependencies to the given tag
}
