% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy-load-dependencies.R
\name{add_busy_load_deps}
\alias{add_busy_load_deps}
\title{busy-load dependencies utils}
\usage{
add_busy_load_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches busy-load. dependencies to the given tag
}
