% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{add_connection}
\alias{add_connection}
\title{Add connection between 2 blocks}
\usage{
add_connection(con, edges, rv)
}
\arguments{
\item{con}{Edge id. Character.}

\item{edges}{Edges dataframe.}

\item{rv}{Reactivevalues containing connections information.}
}
\value{
A list with new observers and connections reactive values.
}
\description{
Add connection between 2 blocks
}
\keyword{internal}
