% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{add_edge}
\alias{add_edge}
\title{Add a edge}
\usage{
add_edge(id = NULL, from, to, label, vals, create_link = TRUE)
}
\arguments{
\item{id}{Edge id. If create_link is FALSE, id cannot be not NULL.}

\item{from}{Node id.}

\item{to}{Node it.}

\item{label}{Edge label. This is useful to map the existing connected
nodes to the input slots of the receiving node (for instance a join block).}

\item{vals}{Global vals reactive values. Read-write access.}

\item{create_link}{Create a link in the board?}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
