% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{add_node_to_stack}
\alias{add_node_to_stack}
\title{Add node to a group}
\usage{
add_node_to_stack(id, stack_id, color, parent)
}
\arguments{
\item{id}{Node id.}

\item{stack_id}{Stack id.}

\item{color}{Stack color.}

\item{parent}{Global reactive values to update data.}
}
\description{
Given a node, update its data to add it to a group (color, group, ...)
}
\keyword{internal}
