% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prismjs-dependencies.R
\name{add_prismjs_deps}
\alias{add_prismjs_deps}
\title{prismjs dependencies utils}
\usage{
add_prismjs_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches prismjs. dependencies to the given tag
}
