% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-add_rm_blocks.R
\name{add_rm_block_server}
\alias{add_rm_block_server}
\alias{add_rm_block_ui}
\title{Add/remove block module}
\usage{
add_rm_block_server(id, board, update, ...)

add_rm_block_ui(id, board)
}
\arguments{
\item{id}{Namespace ID.}

\item{board}{The initial \code{board} object}

\item{update}{Reactive value object to initiate board updates.}

\item{...}{Extra arguments passed from parent scope. Useful to communicate
between plugins and surface information at the top level (for testing ...).}
}
\value{
NULL.
}
\description{
Customizable logic for adding/removing blocks to the board.
}
