% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-stacks.R
\name{add_rm_stack_server}
\alias{add_rm_stack_server}
\alias{add_rm_stack_ui}
\title{Add/remove block stacks module}
\usage{
add_rm_stack_server(id, board, update, ...)

add_rm_stack_ui(id, board)
}
\arguments{
\item{id}{Namespace ID.}

\item{board}{The initial \code{board} object}

\item{update}{Reactive value object to initiate board updates.}

\item{...}{Extra arguments passed from parent scope.}
}
\value{
A reactive value that evaluates to \code{NULL} or a list with components
\code{add} and \code{rm}, where \code{add} is either \code{NULL} or a \code{stacks} object and \code{rm}
is either \code{NULL} or a character vector of link IDs.
}
\description{
Customizable logic for adding/removing stacks grouping blocks together on
the board.
}
