% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-links.R
\name{gen_add_rm_link_server}
\alias{gen_add_rm_link_server}
\alias{add_rm_link_ui}
\title{Add/remove block links module}
\usage{
gen_add_rm_link_server(context_menu)

add_rm_link_ui(id, board)
}
\arguments{
\item{context_menu}{Context menu.}

\item{board}{The initial \code{board} object}
}
\value{
NULL.
}
\description{
Customizable logic for adding/removing links between blocks on the
board.
}
