% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{addStack}
\alias{addStack}
\alias{addStackUI}
\alias{add_stack_server}
\title{Add Stack}
\usage{
addStackUI(id, target)

add_stack_server(id)
}
\arguments{
\item{id}{The id of button stack.}

\item{target}{Selector of target element where stacks can be dropped.}
}
\description{
Create an "add stack" button.
}
