% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ai-tools.R
\name{create_block_tool_factory}
\alias{create_block_tool_factory}
\alias{create_block_names_tool}
\alias{create_remove_block_tool}
\alias{create_add_stack_tool}
\alias{create_stackable_blocks_tool}
\alias{create_get_stack_ids_tool}
\alias{create_add_block_to_stack_tool}
\alias{create_add_to_dash_tool}
\alias{create_remove_from_dash_tool}
\title{Create a block tool factory}
\usage{
create_block_tool_factory(provider, app_request, board, parent, session)

create_block_names_tool(provider)

create_remove_block_tool(provider, app_request, board, parent, session)

create_add_stack_tool(
  provider,
  stackable_blocks,
  app_request,
  board,
  parent,
  session
)

create_stackable_blocks_tool(provider, stackable_blocks)

create_get_stack_ids_tool(provider, board)

create_add_block_to_stack_tool(
  provider,
  stackable_blocks,
  app_request,
  board,
  parent,
  session
)

create_add_to_dash_tool(provider, app_request, board, parent, session)

create_remove_from_dash_tool(provider, app_request, board, parent, session)
}
\arguments{
\item{provider}{AI provider object.}

\item{app_request}{Reactive value containing the tool answer. Maybe useful
for post processing.}

\item{board}{Board object. Not used yet but may be useful.}

\item{parent}{Parent reactive values object.}

\item{session}{Shiny session object.}

\item{stackable_blocks}{Reactive expression returning a vector of blocks
that can be stacked.}
}
\description{
create a tool for a specific block type constructor.
}
