% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-ai-chat.R
\name{chat_mod_ui}
\alias{chat_mod_ui}
\alias{chat_mod_srv}
\title{Chat module UI}
\usage{
chat_mod_ui(id, board, ...)

chat_mod_srv(id = "chat", board, update, session, parent, ...)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board reactiveValues. Read-only.}

\item{...}{Extra parameters.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{session}{Shiny session.}

\item{parent}{Parent global reactiveValues.}
}
\description{
Chat module UI

Chat module server
}
