% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{apply_validation}
\alias{apply_validation}
\title{Apply block validation to network elements}
\usage{
apply_validation(message, id, vals, session)
}
\arguments{
\item{message}{Message.}

\item{id}{Node id.}

\item{vals}{Global reactive values. Read-write.}

\item{session}{Shiny session object.}
}
\description{
Block validation is made by the backend
this function only updates the node color
based on the valid status.
}
