% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{blk_icon}
\alias{blk_icon}
\title{Create block icon}
\usage{
blk_icon(category)
}
\arguments{
\item{category}{Block category. See \link[blockr.core]{available_blocks}.}
}
\description{
Blocks are categorized. This function
returns the corresponding icon for each category.
It may be used in different places like in the nodes
or in the contextual menu (scoutbar) ...
}
\keyword{internal}
