% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block_ui.dag_board}
\alias{block_ui.dag_board}
\alias{insert_block_ui.dag_board}
\alias{remove_block_ui.dag_board}
\title{Block custom UI}
\usage{
\method{block_ui}{dag_board}(id, x, block = NULL, ...)

\method{insert_block_ui}{dag_board}(id, x, blocks = NULL, ...)

\method{remove_block_ui}{dag_board}(id, x, blocks = NULL, ...)
}
\arguments{
\item{id}{Block module id.}

\item{x}{Board object.}

\item{block}{Block to generate the UI for.}

\item{...}{Generic consistency.}
}
\description{
Block custom UI
}
