% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{blockList}
\alias{blockList}
\alias{blockListUI}
\alias{block_list_server}
\title{Block List Module}
\usage{
blockListUI(id, headers = shiny::h5)

block_list_server(id)
}
\arguments{
\item{id}{ID of module.}

\item{headers}{Function to use for headers.}
}
\value{
\link[shiny:reactiveValues]{shiny::reactiveValues} with \code{block} to add where, .
as well as \code{error} to display.
}
\description{
Block List Module
}
