% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-serialize.R
\name{blockr_ser.custom_board}
\alias{blockr_ser.custom_board}
\alias{blockr_deser.custom_board}
\title{Ser/deser board}
\usage{
\method{blockr_ser}{custom_board}(x, blocks = NULL, ...)

\method{blockr_deser}{custom_board}(x, data, ...)
}
\arguments{
\item{blocks}{Board blocks.}

\item{...}{Generic consistency.}
}
\description{
Ser/deser board
}
