% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{add_block_to_grid}
\alias{add_block_to_grid}
\alias{remove_block_from_grid}
\alias{manage_board_grid}
\title{Add block to grid}
\usage{
add_block_to_grid(blk, rv, session)

remove_block_from_grid(blk, session)

manage_board_grid(rv, dashboard_blocks, session)
}
\arguments{
\item{blk}{Block object.}

\item{rv}{Reactive values.}

\item{session}{Shiny session object.}

\item{dashboard_blocks}{List of blocks that should be in the grid.}
}
\description{
Add block to grid

Remove block from grid

Manage board grid
}
