% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dashboard.R, R/utils-module.R
\name{new_dashboard_module}
\alias{new_dashboard_module}
\alias{new_board_module}
\title{Create a board module}
\usage{
new_dashboard_module(id = "dashboard", title = "Dashboard")

new_board_module(
  ui,
  server,
  id,
  title,
  context_menu = list(),
  position = NULL,
  options = new_board_options(),
  class = character()
)
}
\arguments{
\item{id, title}{Module ID and title}

\item{ui, server}{UI and server functions}

\item{context_menu}{List of context menu entries}

\item{position}{Panel position}

\item{options}{Module options (see \code{\link[blockr.core:new_board_options]{blockr.core::new_board_options()}}).}

\item{class}{(Optional) additional class(es)}
}
\description{
Extend a DAG board by adding modules.
}
