% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{board_grid}
\alias{board_grid}
\alias{board_properties}
\title{Grid sidebar}
\usage{
board_grid(..., id, width = "75\%", title = "Dashboard")

board_properties(..., id, width = "40\%", title = "Block properties")
}
\arguments{
\item{...}{Sidebar content.}

\item{id}{Sidebar id.}

\item{width}{Sidebar width.}

\item{title}{Sidebar title}
}
\description{
Default grid sidebar.

Default block sidebar.
}
\keyword{internal}
