% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{board_filename}
\alias{board_filename}
\title{Create board filename}
\usage{
board_filename(rv)
}
\arguments{
\item{rv}{Internal reactiveValues for read-only usage.}
}
\description{
Create board filename
}
\keyword{internal}
