% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{board_ui.dash_board}
\alias{board_ui.dash_board}
\title{Custom board UI}
\usage{
\method{board_ui}{dash_board}(id, x, plugins = list(), ...)
}
\arguments{
\item{id}{Namespace ID.}

\item{x}{Board.}

\item{plugins}{UI for board plugins.}

\item{...}{Generic consistency.}
}
\description{
Custom board UI
}
