% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{build_block}
\alias{build_block}
\title{Construct Block}
\usage{
build_block(x, type)
}
\arguments{
\item{x}{Object to construct block from.}

\item{type}{Type of block.}
}
\description{
Construct block object.
}
\keyword{internal}
