% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-network.R
\name{can_connect}
\alias{can_connect}
\title{Check whether the node can receive connection}
\usage{
can_connect(x, con, vals)
}
\arguments{
\item{x}{Block object.}

\item{con}{Connection. List containing from and to data.}

\item{vals}{Reactive values}
}
\description{
Check whether the node can receive connection
}
