% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{check_connections}
\alias{check_connections}
\title{Check node connection}
\usage{
check_connections(x, vals)
}
\arguments{
\item{x}{block.}

\item{vals}{Reactive values.}
}
\description{
Check node connection
}
