% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{cleanup_node}
\alias{cleanup_node}
\title{Remove a node and associated edges}
\usage{
cleanup_node(selected, vals, rv, session)
}
\arguments{
\item{selected}{UID (character string) of node to remove.}

\item{vals}{Reactive values with dataframe representing nodes data.}

\item{rv}{Board reactive values. Read-only.}

\item{session}{Shiny session object.}
}
\description{
Combine \link{remove_node} with \link{remove_edge}.
}
\keyword{internal}
