% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{create_edges_data_from_links}
\alias{create_edges_data_from_links}
\alias{create_nodes_data_from_blocks}
\alias{create_combos_data_from_stacks}
\alias{cold_start}
\title{Create network data from board}
\usage{
create_edges_data_from_links(links)

create_nodes_data_from_blocks(blocks, stacks)

create_combos_data_from_stacks(stacks, vals, colors)

cold_start(vals, rv, parent, session)
}
\arguments{
\item{links}{Board links.}

\item{blocks}{Board blocks.}

\item{stacks}{Board stacks.}

\item{vals}{Module internal reactive values.}

\item{colors}{Stacks colors. Internal.}

\item{rv}{Board reactive values. Read-only}

\item{parent}{Global app reactive values.}

\item{session}{Shiny session}
}
\description{
That's different from \link{restore_network}, as the
latter restore a network from a JSON compatible structure.
Here we need to re-create all the JSON from the board blocks,
links and stacks.
}
\keyword{internal}
