% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{collapse_container}
\alias{collapse_container}
\title{Create a Bootstrap collapse container.}
\usage{
collapse_container(id, ...)
}
\arguments{
\item{id}{Unique id.}

\item{...}{Content.}
}
\value{
A HTML tag object representing the dropdown element.
}
\description{
Creates a collapsible element.
}
\keyword{internal}
