% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{init_connection}
\alias{init_connection}
\alias{add_connection}
\alias{remove_connection}
\title{Init connections for a block}
\usage{
init_connection(blk)

add_connection(con, edges, rv)

remove_connection(con, rv)
}
\arguments{
\item{blk}{Block object.}

\item{con}{Edge id. Character.}

\item{edges}{Edges dataframe.}

\item{rv}{Reactivevalues containing connections information.}
}
\value{
A list with new observers and connections reactive values.
}
\description{
Init connections for a block

Add connection between 2 blocks

Remove connection between 2 blocks
}
