% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-menu.R
\name{new_context_menu_entry}
\alias{new_context_menu_entry}
\alias{context_menu_items}
\alias{context_menu_items.board_module}
\alias{context_menu_items.list}
\alias{context_menu_items.dag_board}
\title{Create a context menu entry}
\usage{
new_context_menu_entry(
  name,
  js,
  action = NULL,
  condition = TRUE,
  id = tolower(gsub(" +", "_", name))
)

context_menu_items(x)

\method{context_menu_items}{board_module}(x)

\method{context_menu_items}{list}(x)

\method{context_menu_items}{dag_board}(x)
}
\arguments{
\item{name}{Name of the context menu entry.}

\item{js}{JavaScript code to execute when the entry is selected.}

\item{action}{Action to perform when the entry is selected.}

\item{condition}{Condition to determine if the entry should be shown.}

\item{id}{Unique identifier for the context menu entry.
Infered from \code{name} if not provided.}

\item{x}{Object}
}
\description{
Adds a new entry to the context menu of a board.
}
