% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links-g6.R
\name{create_g6_node}
\alias{create_g6_node}
\title{Create a new node it to the network}
\usage{
create_g6_node(new, vals, rv, validate = TRUE, obs, session)
}
\arguments{
\item{new}{New block to add. A block object.}

\item{vals}{Global reactive values. To communicate between modules.}

\item{rv}{Board reactive values. Read-only.}

\item{validate}{Whether to register validation observers for the node.
Default to TRUE.}

\item{obs}{Observer list.}

\item{session}{Shiny session object.}
}
\description{
The rv are updated and the graph
proxy is also updated. We handle either adding new node
or append new node to an existing one. The later case,
involves some edge creation (connection). No validation is
required in practice as a node can theoretically feed
as many children nodes as required.
}
\keyword{internal}
