% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{create_in_con_ui.block}
\alias{create_in_con_ui.block}
\title{Create UI for incoming connection management}
\usage{
\method{create_in_con_ui}{block}(x, other_nodes, ns, rv...)
}
\arguments{
\item{x}{Block.}

\item{other_nodes}{Other node dataframe.}

\item{ns}{Namespace.}

\item{rv}{Reactive values passed between modules.}

\item{...}{Other parameters.}
}
\description{
Create UI for incoming connection management
}
