% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{create_network_widget}
\alias{create_network_widget}
\title{Create network widget}
\usage{
create_network_widget(
  init_nodes = data.frame(),
  init_edges = data.frame(),
  ns,
  height = "100vh",
  width = "100\%",
  interactions = default_network_interactions(),
  options = default_network_options(),
  edges = default_edges_options(),
  physics = default_network_physics(),
  events = default_network_events(ns)
)
}
\arguments{
\item{init_nodes}{Initial set of nodes.}

\item{init_edges}{Initial set of edges.}

\item{ns}{Namespace.}

\item{height}{Network height.}

\item{width}{Network width.}

\item{interactions}{See \link{default_network_interactions}.}

\item{options}{See \link{default_network_options}.}

\item{edges}{See \link{default_edges_options}.}

\item{physics}{See \link{default_network_physics}.}

\item{events}{See \link{default_network_events}.}
}
\value{
A visNetwork object.
}
\description{
Network is populated via a proxy. This function initialises an empty
network with the right setup.
}
\keyword{internal}
