% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{create_node}
\alias{create_node}
\title{Create a new node it to the network}
\usage{
create_node(new, vals, rv, session)
}
\arguments{
\item{new}{New block to add.}

\item{vals}{Local reactive values.}

\item{rv}{Global parent reactive values.}

\item{session}{Shiny session object.}
}
\description{
This is different from \link{add_node}, the later
is just involved to add a row in a dataframe.
The rv are updated and the graph
proxy is also updated. We handle either adding new node
or append new node to an existing one. The later case,
involves some edge creation (connection). No validation is
required in practice as a node can theoretically feed
as many children nodes as required.
}
