% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{create_stack}
\alias{create_stack}
\title{Dynamically group nodes}
\usage{
create_stack(vals, rv, parent, session)
}
\arguments{
\item{vals}{Local scope (links module) reactive values.}

\item{rv}{Board reactive values.}

\item{parent}{Global scope (entire app) reactive values.}

\item{session}{Shiny session object.}
}
\description{
Given a set of selected nodes, add them to a unique group
and apply unique color and labels.
}
\keyword{internal}
