% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createBlock}
\alias{createBlock}
\alias{createBlockUI}
\alias{create_block_server}
\title{Create Block}
\usage{
createBlockUI(id)

create_block_server(id)
}
\arguments{
\item{id}{ID of module.}
}
\description{
Module to create a custom block.
}
