% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block_ui.custom_board}
\alias{block_ui.custom_board}
\alias{remove_block_ui.custom_board}
\title{Block custom UI}
\usage{
\method{block_ui}{custom_board}(id, x, blocks = NULL, ...)

\method{remove_block_ui}{custom_board}(id, x, blocks = NULL, ...)
}
\arguments{
\item{id}{Block module id.}

\item{x}{Block object.}

\item{blocks}{(Additional) blocks (or IDs) for which to generate the UI}
}
\description{
Block custom UI
}
