% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{custom_icons}
\alias{custom_icons}
\title{Needed to be able to use custom SVG icons in G6 toolbar
Icons are taken from fontawesome}
\usage{
custom_icons()
}
\description{
Needed to be able to use custom SVG icons in G6 toolbar
Icons are taken from fontawesome
}
\keyword{internal}
