% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboard_ui}
\alias{dashboard_ui}
\alias{dashboard_server}
\title{Dashboard UI}
\usage{
dashboard_ui(id)

dashboard_server(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Dashboard UI

Dashboard server
}
