% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_g6_options}
\alias{default_g6_options}
\alias{default_g6_behaviors}
\alias{default_g6_plugins}
\title{Default g6 network options}
\usage{
default_g6_options(graph, ...)

default_g6_behaviors(graph, ..., ns)

default_g6_plugins(graph, ..., ns, path, context_menu)
}
\arguments{
\item{graph}{g6 network instance.}

\item{...}{Extra option parameters.}

\item{ns}{Module namespace.}
}
\description{
Default g6 network options
}
\keyword{internal}
