% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_edges_options}
\alias{default_edges_options}
\title{Default edges options}
\usage{
default_edges_options(...)
}
\arguments{
\item{...}{Extra parameters not in defaults accepted by \link[visNetwork]{visEdges}.}
}
\description{
What options to support
}
\keyword{internal}
