% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_network}
\alias{default_network}
\title{Build configuration parameter list}
\usage{
default_network(func, defaults, ...)
}
\arguments{
\item{func}{Function applied.}

\item{defaults}{List of default parameters.}

\item{...}{Extra parameters not in defaults accepted by func.}
}
\description{
Used by all network utilities for default options
}
\keyword{internal}
