% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_network_events}
\alias{default_network_events}
\title{Default events options}
\usage{
default_network_events(ns, ...)
}
\arguments{
\item{...}{Extra parameters not in defaults accepted by \link[visNetwork]{visEvents}.}
}
\description{
What options to support
}
\keyword{internal}
