% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_network_interactions}
\alias{default_network_interactions}
\title{Default network interactions}
\usage{
default_network_interactions(...)
}
\arguments{
\item{...}{Extra parameters not in defaults accepted by \link[visNetwork]{visInteraction}.}
}
\description{
What interaction to support
}
\keyword{internal}
