% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_network_options}
\alias{default_network_options}
\title{Default network options}
\usage{
default_network_options(...)
}
\arguments{
\item{...}{Extra parameters not in defaults accepted by \link[visNetwork]{visOptions}.}
}
\description{
What options to support
}
\keyword{internal}
