% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{default_network_physics}
\alias{default_network_physics}
\title{Default physics options}
\usage{
default_network_physics(...)
}
\arguments{
\item{...}{Extra parameters not in defaults accepted by \link[visNetwork]{visPhysics}.}
}
\description{
What options to support
}
\keyword{internal}
