% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{define_conlabel}
\alias{define_conlabel}
\title{Define connection label}
\usage{
define_conlabel(x, target, rv)
}
\arguments{
\item{x}{Block object.}

\item{target}{Connection target id.}

\item{rv}{Board reactive values}
}
\description{
Define connection label
}
