% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency}
\alias{dependency}
\title{Dependencies}
\usage{
dependency(dep)
}
\arguments{
\item{dep}{Name of dependency, generally corresponding to a file name.}
}
\description{
Attache dependncies.
}
\keyword{internal}
