% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock.R
\name{dock_ui}
\alias{dock_ui}
\alias{dock_server}
\title{Dashboard dock UI}
\usage{
dock_ui(id)

dock_server(board, update, parent, ...)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board reactiveValues. Read-only.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{parent}{Parent global reactiveValues.}

\item{...}{Extra parameters.}
}
\description{
Dashboard dock UI

Dashboard dock server
}
