% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{dropdown_button}
\alias{dropdown_button}
\title{Create a Bootstrap dropdown}
\usage{
dropdown_button(..., icon)
}
\arguments{
\item{...}{Content.}

\item{icon}{Icon.}
}
\value{
A HTML tag object representing the dropdown element.
}
\description{
Creates a dropdown menu.
}
