% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-code.R
\name{generate_code_ui}
\alias{generate_code_ui}
\title{Code generation module}
\usage{
generate_code_ui(id, board)
}
\arguments{
\item{id}{Module id.}

\item{board}{The initial \code{board} object}
}
\description{
Generate reproducible code from a board.
}
