% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{get_block_registry}
\alias{get_block_registry}
\title{Get block info in registry}
\usage{
get_block_registry(x)
}
\arguments{
\item{x}{Block object}
}
\description{
Get block info in registry
}
\keyword{internal}
