% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{get_blocks_state}
\alias{get_blocks_state}
\title{Get the state of a block}
\usage{
get_blocks_state(rv)
}
\arguments{
\item{rv}{Board reactiveValues for read-only usage.}
}
\description{
Get the state of a block
}
\keyword{internal}
