% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_ui}
\alias{grid_ui}
\alias{grid_server}
\title{Dashboard grid UI}
\usage{
grid_ui(id)

grid_server(board, update, parent, ...)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board reactiveValues. Read-only.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{parent}{Parent global reactiveValues.}

\item{...}{Extra parameters.}
}
\description{
Dashboard grid UI

Dashboard grid server
}
