% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{handle_dashboard_zoom}
\alias{handle_dashboard_zoom}
\title{Update dashboard zoom on the client}
\usage{
handle_dashboard_zoom(session)
}
\arguments{
\item{session}{Shiny session object}
}
\description{
This allows to set different zoom level to handle more
crowded dashboards.
}
\keyword{internal}
