% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{init_dashboard_state}
\alias{init_dashboard_state}
\alias{init_dashboard_state.dock_board}
\alias{init_dashboard_state.grid_board}
\title{Init dashboard state generic}
\usage{
init_dashboard_state(board, blocks, vals)

init_dashboard_state.dock_board(board, blocks, vals)

init_dashboard_state.grid_board(board, blocks, vals)
}
\arguments{
\item{blocks}{Board blocks.}

\item{vals}{Local reactive values.}
}
\description{
Whenever a new block is created or some blocks
are removed, we update the mapping to know which
block should be in the dashboard, so that \link{manage_dashboard}
knows what to do.
}
