% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{init_block}
\alias{init_block}
\title{Init block ui and server module}
\usage{
init_block(blk, rv, ns)
}
\arguments{
\item{blk}{Block object.}

\item{rv}{Reactivevalues containing connections information.}

\item{ns}{Namespace.}
}
\description{
Init block ui and server module
}
\keyword{internal}
