% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{init_block_server}
\alias{init_block_server}
\title{Init block server module}
\usage{
init_block_server(blk, rv)
}
\arguments{
\item{blk}{Block object.}

\item{rv}{Reactivevalues containing connections information.}
}
\description{
Init block server module
}
\keyword{internal}
