% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{init_connection}
\alias{init_connection}
\title{Init connections for a block}
\usage{
init_connection(blk)
}
\arguments{
\item{blk}{Block object.}
}
\description{
Init connections for a block
}
\keyword{internal}
