% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{is_input_slot_connected}
\alias{is_input_slot_connected}
\title{Check whether a block input is connected}
\usage{
is_input_slot_connected(x, slot, vals)
}
\arguments{
\item{x}{Block object.}

\item{slot}{Connection slot.}

\item{vals}{Reactive values passed between modules.}
}
\description{
Check whether a block input is connected
}
