% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-network.R
\name{list_empty_connections}
\alias{list_empty_connections}
\title{List node connections}
\usage{
list_empty_connections(x, con, rv)
}
\arguments{
\item{x}{block.}

\item{con}{Connection information.}

\item{rv}{Reactive values.}
}
\description{
List node connections
}
