% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{make_send_message}
\alias{make_send_message}
\title{Send Factory}
\usage{
make_send_message(prefix, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{prefix}{Prefix to use all messages.}

\item{session}{\link[shiny:session]{shiny::session} to use.}
}
\description{
Makes a function that sends a custom message to the client.
}
\keyword{internal}
