% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{add_dashboard_block}
\alias{add_dashboard_block}
\alias{add_dashboard_block.dock_board}
\alias{remove_dashboard_block}
\alias{remove_dashboard_block.dock_board}
\alias{manage_dashboard}
\alias{manage_dashboard.dock_board}
\title{Add block to dashboard generic}
\usage{
add_dashboard_block(board, id, vals, session)

\method{add_dashboard_block}{dock_board}(board, id, vals, session)

remove_dashboard_block(board, id, session)

\method{remove_dashboard_block}{dock_board}(board, id, session)

manage_dashboard(board, mode, vals, session)

\method{manage_dashboard}{dock_board}(board, mode, vals, session)
}
\arguments{
\item{board}{Board.}

\item{id}{Block id.}

\item{vals}{Local reactive values.}

\item{session}{Shiny session object.}

\item{mode}{App mode.}
}
\description{
Add block to dashboard generic

Add block to dock method

Remove block from dashboard generic

Remove block from dock method

Manage dashboard generic

Manage board dock method
}
