% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ai-tools.R
\name{manage_chat}
\alias{manage_chat}
\title{Create a chat server callbacks}
\usage{
manage_chat(provider, parent, session)
}
\arguments{
\item{provider}{AI provider object. See \url{setup_chat_provider}.}

\item{parent}{Parent global reactive values object.}

\item{session}{Shiny session object.}
}
\description{
Create a chat server callbacks
}
