% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{blks_metadata}
\alias{blks_metadata}
\alias{blk_color}
\title{Get metadata for blocks}
\usage{
blks_metadata(blocks)

blk_color(category)
}
\arguments{
\item{blocks}{Blocks passed as \code{blocks} or \code{block} object}

\item{category}{Block category}
}
\description{
Get metadata for blocks

Get block color based on category
}
