% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_ui}
\alias{network_ui}
\alias{network_server}
\title{Network UI}
\usage{
network_ui(id, debug = TRUE)

network_server(id, vals, debug = TRUE)
}
\arguments{
\item{id}{Module id.}

\item{debug}{Whether to display debug information.}

\item{vals}{Reactive values of parent module.}
}
\description{
Network UI

Network server module
}
