% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-blocks.R
\name{new_readcsv_block}
\alias{new_readcsv_block}
\title{Data upload block constructor}
\usage{
new_readcsv_block(sep = ",", quote = "\\"", ...)
}
\arguments{
\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
In order to make user data available to blockr, this block provides file-
upload functionality.
}
