% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-blocks.R
\name{new_readxpt_block}
\alias{new_readxpt_block}
\title{XPT data upload block constructor}
\usage{
new_readxpt_block(col_select = NULL, skip = 0, n_max = Inf, ...)
}
\arguments{
\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
In order to make user data available to blockr, this block provides xpt file-
upload functionality.
}
