% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_app_state}
\alias{process_app_state}
\title{Useful for shinytest2
Pre-process reactiveValues results}
\usage{
process_app_state(state)
}
\description{
Useful for shinytest2
Pre-process reactiveValues results
}
\keyword{internal}
