% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{process_grid_content}
\alias{process_grid_content}
\title{Process grid layout}
\usage{
process_grid_content(rv, grid_layout, dashboard_blocks)
}
\arguments{
\item{rv}{Reactive values}

\item{grid_layout}{Returned by input$<GRID_ID>_layout.
Contains blocks coordinates, dimensions, ...}

\item{dashboard_blocks}{Blocks that should be in the grid.}
}
\description{
Process grid layout
}
\keyword{internal}
