% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{register_add_to_grid_obs}
\alias{register_add_to_grid_obs}
\title{Register observers related to the add to grid switches}
\usage{
register_add_to_grid_obs(blocks_ids, parent, rv, obs, session)
}
\arguments{
\item{parent}{Parent scope reactive values.}

\item{rv}{Board reactive values.}

\item{obs}{Observers list.}

\item{session}{Shiny session object.}

\item{block_ids}{Board block ids.}
}
\description{
Observer to maintain the state between the 2 switches +
an observer to handle serialisation/restoration.
}
\keyword{internal}
