% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{register_node_stack_link}
\alias{register_node_stack_link}
\title{Register block stack bind/unbind}
\usage{
register_node_stack_link(id, rv, vals, session)
}
\arguments{
\item{id}{Block id for which to register the validation.}

\item{rv}{Board reactive values. Read-only.}

\item{vals}{Global reactive values. Read-write.}

\item{session}{Shiny session object.}
}
\description{
For each block we register an observer that
captures only the messages related to this block validation
status.
}
\keyword{internal}
