% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{remove_connection}
\alias{remove_connection}
\title{Remove connection between 2 blocks}
\usage{
remove_connection(con, rv)
}
\arguments{
\item{con}{Edge id to remove. Character.}

\item{rv}{Reactivevalues containing connections information.}
}
\description{
Remove connection between 2 blocks
}
\keyword{internal}
