% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{remove_node}
\alias{remove_node}
\title{Remove a node}
\usage{
remove_node(selected, nodes)
}
\arguments{
\item{selected}{UID (character string) of node to remove.}

\item{nodes}{dataframe representing nodes data.}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
