% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{restore_board}
\alias{restore_board}
\title{Restore board from snapshot}
\usage{
restore_board(path, res, parent)
}
\arguments{
\item{path}{JSON snapshot path.}

\item{res}{reactiveVal containing the module returned value.}

\item{parent}{Parent reactiveValues to communicate to other modules.}
}
\description{
Restore board from snapshot
}
\keyword{internal}
