% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{restore_grid}
\alias{restore_grid}
\title{Restore grid state from board state}
\usage{
restore_grid(blocks, blocks_ns, vals, parent)
}
\arguments{
\item{blocks}{Board block objects.}

\item{blocks_ns}{Blocks namespace.}

\item{vals}{Local reactive values.}

\item{parent}{Parent reactive values.
Contains blocks coordinates, dimensions, ...}
}
\description{
Restore grid state from board state
}
\keyword{internal}
