% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{restore_network}
\alias{restore_network}
\title{Restore network from saved snapshot}
\usage{
restore_network(rv, vals, session)
}
\arguments{
\item{rv}{Board internal reactive values. Read-only}

\item{vals}{Global vals reactive values. Read-write access.}

\item{session}{Shiny session object}
}
\value{
A reactiveValues object.
}
\description{
Network is updated via a proxy.
}
\keyword{internal}
