% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{board_filename}
\alias{board_filename}
\alias{write_board_to_disk}
\title{Create board filename}
\usage{
board_filename(rv)

write_board_to_disk(rv, parent, session)
}
\arguments{
\item{rv}{Internal reactiveValues for read-only usage.}

\item{parent}{Parent reactiveValues to communicate to other modules.}

\item{session}{Shiny session object.}
}
\description{
Create board filename

Save board to disk
}
\keyword{internal}
