% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-serialize.R
\name{ser_deser_server}
\alias{ser_deser_server}
\alias{ser_deser_ui}
\title{Serialization module}
\usage{
ser_deser_server(id, rv, ...)

ser_deser_ui(id, board)
}
\arguments{
\item{id}{module ID.}

\item{rv}{Reactive values object}

\item{...}{Extra arguments passed from parent scope}

\item{board}{The initial \code{board} object}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object that evaluates to \code{NULL} or a
\code{board} obejct.
}
\description{
Object (de)serialization in a board server context.
}
