% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{setup_remove_elements_kbd}
\alias{setup_remove_elements_kbd}
\title{Remove element from keyboard}
\usage{
setup_remove_elements_kbd(session)
}
\description{
Register event to remove selected elements.
Can be edge or nodes. At the moment only 1 edge
can be removed at a time. Multiple nodes can be
selected and removed at the same time.
}
\keyword{internal}
