% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{show_node_menu}
\alias{show_node_menu}
\title{Create and show node menu}
\usage{
show_node_menu(value, session)
}
\arguments{
\item{value}{Initial state of the grid switch. Depends
on the value set in the grid parent module.}

\item{session}{Shiny session object}
}
\description{
Node menu contains shortcut to some actions
like adding a node to a grid...
}
\keyword{internal}
