% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{show_stack_actions}
\alias{show_stack_actions}
\title{Show stack actions}
\usage{
show_stack_actions(selected, rv, parent, session)
}
\arguments{
\item{selected}{Set of selected nodes.}

\item{rv}{Reactive values containing board elements. Read-only.}

\item{parent}{Global reactive values.}

\item{session}{Shiny session object.}
}
\description{
A modal window triggered when multiple nodes are selected, for instance.
So far we support adding a stack and removing multiple blocks from a stack.
}
\keyword{internal}
