% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{snapshot_board}
\alias{snapshot_board}
\title{Capture board snapshot}
\usage{
snapshot_board(vals, rv, parent)
}
\arguments{
\item{vals}{Local reactiveValues.}

\item{rv}{Internal reactiveValues for read-only usage.}

\item{parent}{Parent reactiveValues to communicate to other modules.}
}
\description{
This is used to autosnapshot the board.
}
\keyword{internal}
