% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links-g6.R
\name{stack_g6_nodes}
\alias{stack_g6_nodes}
\alias{unstack_g6_nodes}
\title{Dynamically group nodes}
\usage{
stack_g6_nodes(stack_id = NULL, nodes = NULL, vals, rv, parent, session)

unstack_g6_nodes(vals, parent, session)
}
\arguments{
\item{stack_id}{Stack id to attach nodes to.}

\item{nodes}{Vector of node ids to stack.}

\item{vals}{Local scope (links module) reactive values.}

\item{rv}{Board reactive values.}

\item{parent}{Global scope (entire app) reactive values.}

\item{session}{Shiny session object.}
}
\description{
Given a set of selected nodes, add them to a unique group
and apply unique color and labels.

Useful when removing a stack.
}
\keyword{internal}
