% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{stack_nodes}
\alias{stack_nodes}
\title{Dynamically group nodes}
\usage{
stack_nodes(stack_id, color, vals, rv, parent, session)
}
\arguments{
\item{stack_id}{Stack unique id.}

\item{color}{Stack color. Given by color picker input.}

\item{vals}{Local scope (links module) reactive values.}

\item{rv}{Board reactive values.}

\item{parent}{Global scope (entire app) reactive values.}

\item{session}{Shiny session object.}
}
\description{
Given a set of selected nodes, add them to a unique group
and apply unique color and labels.
}
\details{
This function must be called after \link{trigger_create_stack}.
}
\keyword{internal}
