% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stacksArea}
\alias{stacksArea}
\title{Stack Area}
\usage{
stacksArea(id = NULL, class = "", style = "", ...)
}
\arguments{
\item{id}{The id of stack area.}

\item{class}{Additional class to add to stack area.}

\item{style}{Additional style to add to stack area.}

\item{...}{Additional \strong{data} attributes to add to stack area.}
}
\description{
Area where to drop stacks.
}
