% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ai-tools.R
\name{system_prompt}
\alias{system_prompt}
\title{Utility functions to create AI tools for blockr.ui}
\usage{
system_prompt()
}
\description{
Utility functions to create AI tools for blockr.ui
}
\keyword{internal}
