% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\title{Toast}
\usage{
toast(
  ...,
  class = "",
  position = c("top-right", "top-left", "bottom-right", "bottom-left")
)
}
\arguments{
\item{...}{Content, generally \code{\link[=toastHeader]{toastHeader()}}, \code{\link[=toastBody]{toastBody()}}.}

\item{class}{Any additional classes.}

\item{position}{Position of the toast.}
}
\description{
Toast notifications.
}
