% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{toggle_undo_redo}
\alias{toggle_undo_redo}
\title{Toggle undo/redo}
\usage{
toggle_undo_redo(vals, parent)
}
\arguments{
\item{vals}{Local module reactive Values.}

\item{parent}{Global reactive Values.}
}
\description{
Toggle state of undo/redo buttons.
}
\keyword{internal}
