% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{trigger_create_stack}
\alias{trigger_create_stack}
\title{Signal for stack creation}
\usage{
trigger_create_stack(selected, parent)
}
\arguments{
\item{selected}{Selected node id.}

\item{parent}{Global reactive values to update data.}
}
\description{
Given a selected node, find whether a stack can be created.
If succesful, the stack is created from the stacks plugin.
}
\keyword{internal}
