% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{trigger_remove_stack}
\alias{trigger_remove_stack}
\title{Signal for stack removal}
\usage{
trigger_remove_stack(selected, parent)
}
\arguments{
\item{selected}{Selected node id.}

\item{parent}{Global reactive values to update data.}
}
\description{
Given a selected node, find whether a stack can be removed.
If successful, then the parent$removed_stack will be captured
in the stacks plugin.
}
\keyword{internal}
