% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{unstack_nodes}
\alias{unstack_nodes}
\title{Dynamically ungroup nodes}
\usage{
unstack_nodes(vals, rv, parent, session)
}
\arguments{
\item{vals}{Local scope (links module) reactive values.}

\item{rv}{Board reactive values.}

\item{parent}{Global scope (entire app) reactive values.}

\item{session}{Shiny session object.}
}
\description{
Given a set of selected nodes,
ungroup them by removing the associated stack.
}
\details{
This function must be called after \link{trigger_remove_stack}.
}
\keyword{internal}
