% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{update_dashboard_state}
\alias{update_dashboard_state}
\alias{update_dashboard_state.dock_board}
\title{Update block dashboard state generic}
\usage{
update_dashboard_state(board, selected, value, vals)

\method{update_dashboard_state}{dock_board}(board, selected, value, vals)
}
\arguments{
\item{board}{Board object.}

\item{selected}{Selected block id.}

\item{value}{New value.}

\item{vals}{Local reactive values.}
}
\description{
For a given selected block, toggle its dock presence
based on the specified value.
}
