% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{update_block_dashboard_input}
\alias{update_block_dashboard_input}
\title{Update switch input}
\usage{
update_block_dashboard_input(board, selected, value, vals, session)
}
\arguments{
\item{board}{Board object.}

\item{selected}{Selected block id.}

\item{value}{New value.}

\item{vals}{Local reactive values.}

\item{session}{Shiny session object.}
}
\description{
For a given selected block, update the switch input
according to the block dashboard state (if values are different).
}
\keyword{internal}
