% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{update_block_grid_input}
\alias{update_block_grid_input}
\title{Update switch input}
\usage{
update_block_grid_input(selected, value, vals, session)
}
\arguments{
\item{selected}{Selected block id.}

\item{value}{New value.}

\item{vals}{Local reactive values.}

\item{session}{Shiny session object.}
}
\description{
For a given selected block, update the switch input
according to the block grid state (if values are different).
}
\keyword{internal}
