% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{update_block_grid_state}
\alias{update_block_grid_state}
\title{Update block grid state}
\usage{
update_block_grid_state(selected, value, vals)
}
\arguments{
\item{selected}{Selected block id.}

\item{value}{New value.}

\item{vals}{Local reactive values.}
}
\description{
For a given selected block, toggle its grid presence
based on the specified value.
}
\keyword{internal}
