% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{use_blockr_ui_deps}
\alias{use_blockr_ui_deps}
\title{Insert blockr.ui dependencies}
\usage{
use_blockr_ui_deps(tag)
}
\arguments{
\item{tag}{Tag to attach the dependency to.}
}
\description{
Contains JavaScript necessary for the app to run.
}
\keyword{internal}
