% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{validate_edge_creation}
\alias{validate_edge_creation}
\title{Validate edge creation for a block}
\usage{
validate_edge_creation(target, rv)
}
\arguments{
\item{target}{Target block id}

\item{rv}{Board reactive values. Read-only.}
}
\description{
Validate edge creation for a block
}
\keyword{internal}
