% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{write_board_to_disk}
\alias{write_board_to_disk}
\title{Save board to disk}
\usage{
write_board_to_disk(rv, parent)
}
\arguments{
\item{rv}{Internal reactiveValues for read-only usage.}

\item{parent}{Parent reactiveValues to communicate to other modules.}
}
\description{
Save board to disk
}
\keyword{internal}
