% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{add_edge}
\alias{add_edge}
\title{Add a edge}
\usage{
add_edge(from, to, label, rv)
}
\arguments{
\item{from}{Node id.}

\item{to}{Node it.}

\item{label}{Edge label. This is useful to map the existing connected
nodes to the input slots of the receiving node (for instance a join block).}

\item{rv}{Reactive values, containing elements such as edges data.}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
