% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{add_node}
\alias{add_node}
\title{Add a node}
\usage{
add_node(new, vals, rv)
}
\arguments{
\item{new}{New block to add.}

\item{vals}{Local reactive values.}

\item{rv}{Global reactive values.}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
