% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{add_node}
\alias{add_node}
\title{Add a node}
\usage{
add_node(new, rv)
}
\arguments{
\item{new}{New block to insert. Must be a valid
block registry entry}

\item{rv}{Reactive values with dataframe representing network data.}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
