% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-links.R
\name{add_rm_link_server}
\alias{add_rm_link_server}
\alias{add_rm_link_ui}
\title{Add/remove block links module}
\usage{
add_rm_link_server(id, rv, update, ...)

add_rm_link_ui(id, board)
}
\arguments{
\item{id}{Namespace ID.}

\item{rv}{Reactive values object.}

\item{update}{Reactive value object to initiate board updates.}

\item{...}{Extra arguments passed from parent scope.}

\item{board}{The initial \code{board} object}
}
\value{
NULL.
}
\description{
Customizable logic for adding/removing links between blocks on the
board.
}
