% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{addStack}
\alias{addStack}
\alias{addStackUI}
\alias{add_stack_server}
\title{Add Stack}
\usage{
addStackUI(id, item)

add_stack_server(id, on_select = NULL, on_deselect = NULL)
}
\arguments{
\item{id}{The id of button stack.}

\item{item}{Item to use as draggable.}

\item{on_select, on_deselect}{JavaScript to run when the user selects or deselects a block.}
}
\description{
Create an "add stack" button.
}
