% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{addStack}
\alias{addStack}
\alias{addStackUI}
\alias{add_stack_server}
\alias{add_stack_bind}
\title{Add Stack}
\usage{
addStackUI(
  id,
  content = shiny::span(class = "badge bg-primary fs-5 mx-2", shiny::icon("plus"),
    "Stack"),
  class = "badge bg-primary fs-5 mx-2",
  target = ".stack-target",
  toast_position = c("top-right", "top-left", "bottom-right", "bottom-left")
)

add_stack_server(id, feedback = TRUE, delay = 0L)

add_stack_bind(
  ns = "",
  feedback = TRUE,
  session = shiny::getDefaultReactiveDomain(),
  delay = 0L
)
}
\arguments{
\item{id, ns}{The id of add stack function.}

\item{content}{Content of the draggable item.}

\item{target}{Selector of target element where stacks can be dropped.}

\item{toast_position}{Position of toast, only used if \code{feedback} is \code{TRUE}.}

\item{feedback}{Whether to notify user of errors, warnings, and more.}

\item{delay}{Delay in milliseconds before binding JavaScript.}

\item{session}{Shiny session.}
}
\description{
Create an "add stack" button.
}
