% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{blk_choices}
\alias{blk_choices}
\title{Create block choices for scoutbaR widget}
\usage{
blk_choices()
}
\description{
Utility to populate the scoutbar with block
registry information. Create one page per block category
}
