% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{create_or_show_block_panel}
\alias{create_or_show_block_panel}
\alias{hide_block_panel}
\title{Create/Show a block panel}
\usage{
create_or_show_block_panel(proxy, id, parent)

hide_block_panel(proxy, id)
}
\arguments{
\item{proxy}{Dock proxy.}

\item{id}{Block id to show}

\item{parent}{Parent reactive values.}
}
\description{
If panel does not exist, create it and move the block UI from
offcanvas to the panel container. If it exists, just select it.

Move block from panel to offcanvas-body.
}
