% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{create_or_show_block_panel}
\alias{create_or_show_block_panel}
\alias{hide_block_panel}
\title{Show a block panel}
\usage{
create_or_show_block_panel(id, parent, session)

hide_block_panel(id, session)
}
\arguments{
\item{id}{Block id to show}

\item{parent}{Parent reactive values.}

\item{session}{Shiny session object.}
}
\description{
Move block from offcanvas-body to a panel.

Move block from panel to offcanvas-body.
}
