% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{blockList}
\alias{blockList}
\alias{blockListUI}
\alias{block_list_server}
\alias{block_list_bind}
\title{Block List Module}
\usage{
blockListUI(
  id,
  headers = shiny::h5,
  toast_position = c("top-right", "top-left", "bottom-right", "bottom-left"),
  max_height = "100vh"
)

block_list_server(id, feedback = TRUE, delay = 0L)

block_list_bind(session = shiny::getDefaultReactiveDomain(), delay = 0L)
}
\arguments{
\item{id}{ID of module.}

\item{headers}{Function to use for headers.}

\item{toast_position}{Position of toast, only used if \code{feedback} is \code{TRUE}.}

\item{max_height}{Maximum height of block list.}

\item{feedback}{Whether to notify user of errors, warnings, and more.}

\item{delay}{Delay in milliseconds before binding JavaScript.}

\item{session}{Shiny session.}
}
\value{
A named \code{list} containing the following reactives: \code{error}, \code{started}, and \code{dropped}.
}
\description{
Display a list of draggable blocks.
}
\details{
The list is server-side rendered.
}
