% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{blockList}
\alias{blockList}
\alias{blockListUI}
\alias{block_list_server}
\title{Block List Module}
\usage{
blockListUI(
  id,
  headers = shiny::h5,
  toast_position = c("top-right", "top-left", "bottom-right", "bottom-left")
)

block_list_server(id, feedback = TRUE, delay = 0L)
}
\arguments{
\item{id}{ID of module.}

\item{headers}{Function to use for headers.}

\item{toast_position}{Position of toast, only used if \code{feedback} is \code{TRUE}.}

\item{feedback}{Whether to notify user of errors, warnings, and more.}

\item{delay}{Delay in milliseconds before binding JavaScript.}
}
\value{
\link[shiny:reactiveValues]{shiny::reactiveValues} with \code{block} to add where, .
as well as \code{error} to display.
}
\description{
Block List Module
}
