% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockr.ui-package.R
\docType{package}
\name{blockr.ui-package}
\alias{blockr.ui}
\alias{blockr.ui-package}
\title{blockr.ui: User Interface Components for the 'blockr' Ecosystem}
\description{
User-interface primitives shared across the 'blockr' ecosystem. Designed enhance 'blockr.dock'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr.ui/}
  \item Report bugs at \url{https://github.com/BristolMyersSquibb/blockr.ui/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{david@cynkra.com}

}
\keyword{internal}
