% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.ui-package}
\alias{blockr.ui}
\alias{blockr.ui-package}
\title{blockr.ui: A 'shiny' powered user interface for 'blockr.core'}
\description{
Provides a web-based point and click interface for creating data pipelines and visualizations with 'blockr.core'.
}
\author{
\strong{Maintainer}: David Granjon \email{david@cynkra.com}

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item Christoph Sax \email{christoph@cynkra.com}
}

}
\keyword{internal}
