% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{blockr_ser.dag_board}
\alias{blockr_ser.dag_board}
\alias{blockr_deser.dag_board}
\title{Ser/deser board}
\usage{
\method{blockr_ser}{dag_board}(
  x,
  blocks = NULL,
  options = NULL,
  network = NULL,
  selected = NULL,
  modules = NULL,
  ...
)

\method{blockr_deser}{dag_board}(x, data, ...)
}
\arguments{
\item{x}{Board.}

\item{blocks}{Board blocks.}

\item{options}{Board options.}

\item{network}{visNetwork data.}

\item{selected}{Selected node.}

\item{modules}{Module state.}

\item{...}{Generic consistency.}

\item{data}{Data to restore.}
}
\description{
Ser/deser board
}
