% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{blockWrapper}
\alias{blockWrapper}
\title{Block Wrapper}
\usage{
blockWrapper(blocks, cls)
}
\arguments{
\item{blocks}{List of blocks to wrap.}

\item{cls}{Class to filter}
}
\description{
Creates a UI wrapper for a block.
Used to enable sortable.
}
\keyword{internal}
