% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{add_block_to_grid}
\alias{add_block_to_grid}
\alias{remove_block_from_grid}
\alias{manage_board_grid}
\title{Add block to grid}
\usage{
add_block_to_grid(id, vals, blocks_ns, session)

remove_block_from_grid(id, blocks_ns, session)

manage_board_grid(mode, vals, blocks_ns, session)
}
\arguments{
\item{id}{Block id}

\item{vals}{Local reactive values.}

\item{blocks_ns}{Blocks namespace.}

\item{session}{Shiny session object.}

\item{mode}{App mode.}
}
\description{
Add block to grid

Remove block from grid

Manage board grid
}
