% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{board_burger}
\alias{board_burger}
\alias{board_header}
\title{Board extra actions}
\usage{
board_burger(board_ui)

board_header(id, board_ui)
}
\arguments{
\item{board_ui}{Board ui.}

\item{id}{Board id.}
}
\description{
Extra actions dropdown.

Header layout.
}
\keyword{internal}
