% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{board_header}
\alias{board_header}
\title{Board header}
\usage{
board_header(id, board_ui)
}
\arguments{
\item{id}{Board id.}

\item{board_ui}{Board ui.}
}
\description{
Header layout.
}
\keyword{internal}
