% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-module.R
\name{new_board_module}
\alias{new_board_module}
\alias{new_dashboard_module}
\title{Create a board module}
\usage{
new_board_module(
  ui,
  server,
  id,
  title,
  context_menu = list(),
  position = NULL,
  class = character()
)

new_dashboard_module(id = "dashboard", title = "Dashboard")
}
\arguments{
\item{ui, server}{UI and server functions}

\item{id, title}{Module ID and title}

\item{context_menu}{List of context menu entries}

\item{position}{Panel position}

\item{class}{(Optional) additional class(es)}
}
\description{
Extend a DAG board by adding modules.
}
