% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{dashboard_ui}
\alias{dashboard_ui}
\alias{dashboard_server}
\alias{network_ui}
\alias{network_server}
\alias{board_ui}
\alias{board_server}
\title{The board provides the main ui for blockr2}
\usage{
dashboard_ui(id)

dashboard_server(id)

network_ui(id)

network_server(id)

board_ui(id)

board_server(id)
}
\arguments{
\item{id}{Unique id.}
}
\description{
The board is composed of 2 views pointing to the network module
or the dashboard module.
}
