% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R, R/dashboard.R, R/network.R
\name{board_ui}
\alias{board_ui}
\alias{board_server}
\alias{dashboard_ui}
\alias{dashboard_server}
\alias{network_ui}
\alias{network_server}
\title{The board provides the main ui for blockr2}
\usage{
board_ui(id)

board_server(id)

dashboard_ui(id)

dashboard_server(id, vals, parent_ns)

network_ui(id, debug = TRUE)

network_server(id, vals, debug = TRUE)
}
\arguments{
\item{id}{Unique id.}

\item{vals}{Reactive values of parent module.}

\item{debug}{Print extra information about the network}
}
\description{
The board is composed of 2 views pointing to the network module
or the dashboard module.
}
