% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{options_ui}
\alias{options_ui}
\alias{board_ui.dag_board}
\title{Custom board UI}
\usage{
options_ui(id, x, ...)

\method{board_ui}{dag_board}(id, x, ...)
}
\arguments{
\item{id}{Namespace ID.}

\item{x}{Board.}

\item{...}{Generic consistency.}
}
\description{
Custom board UI
}
