% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{board_ui.board_options}
\alias{board_ui.board_options}
\alias{update_ui.board_options}
\alias{board_ui.dag_board}
\title{Custom board UI}
\usage{
\method{board_ui}{board_options}(id, x, ...)

\method{update_ui}{board_options}(x, session, ...)

\method{board_ui}{dag_board}(id, x, plugins = list(), ...)
}
\arguments{
\item{id}{Namespace ID.}

\item{x}{Board.}

\item{...}{Generic consistency.}

\item{plugins}{UI for board plugins.}
}
\description{
Custom board UI
}
