% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-network.R
\name{can_connect}
\alias{can_connect}
\title{Check whether the node can receive connection}
\usage{
can_connect(x, target, rv)
}
\arguments{
\item{x}{Block object.}

\item{target}{Connection target id.}

\item{rv}{Reactive values}
}
\description{
Check whether the node can receive connection
}
