% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-network.R
\name{check_connections}
\alias{check_connections}
\title{Check node connection}
\usage{
check_connections(x, con, rv)
}
\arguments{
\item{x}{block.}

\item{con}{Connection information.}

\item{rv}{Reactive values.}
}
\description{
Check node connection
}
