% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{check_connections}
\alias{check_connections}
\title{Check node connection}
\usage{
check_connections(x, target, rv)
}
\arguments{
\item{x}{block.}

\item{target}{Connection target id.}

\item{rv}{Board reactive values.}
}
\description{
Check node connection
}
