% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-module.R
\name{new_context_menu_entry}
\alias{new_context_menu_entry}
\title{Create a context menu entry}
\usage{
new_context_menu_entry(
  name,
  js,
  action = NULL,
  condition = TRUE,
  id = tolower(gsub(" +", "_", name))
)
}
\arguments{
\item{name}{Name of the context menu entry.}

\item{js}{JavaScript code to execute when the entry is selected.}

\item{action}{Action to perform when the entry is selected.}

\item{condition}{Condition to determine if the entry should be shown.}

\item{id}{Unique identifier for the context menu entry.
Infered from \code{name} if not provided.}
}
\description{
Adds a new entry to the context menu of a board.
}
