% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{create_edge}
\alias{create_edge}
\title{Create an edge and add it to the network}
\usage{
create_edge(new, vals, rv, session)
}
\arguments{
\item{new}{Edge data. A list like
\code{list(source = "from_node_ID", target = "to_node_ID")}.}

\item{vals}{Global reactive values. Read-write.}

\item{rv}{Board reactive values. Read-only.}

\item{session}{Shiny session object.}
}
\description{
There is a validation layer prior to knowing whether
we can add the edge. Then rv are updated and the graph
proxy is also updated.
}
\keyword{internal}
