% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-network.R
\name{create_edge}
\alias{create_edge}
\title{Create an edge and add it to the network}
\usage{
create_edge(vals, rv, session)
}
\arguments{
\item{vals}{Local reactive values.}

\item{rv}{Parent reactive values.}

\item{session}{Shiny session object.}
}
\description{
This is different from \link{add_edge}, the later
is just involved to add a row in a dataframe.
There is a validation layer prior to knowing whether
we can add the edge. Then rv are updated and the graph
proxy is also updated.
}
