% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dashboard.R
\name{dashboard_ui}
\alias{dashboard_ui}
\alias{dashboard_server}
\title{Dashboard UI}
\usage{
dashboard_ui(id, board, ...)

dashboard_server(id, board, update, session, parent, ...)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board reactiveValues. Read-only.}

\item{...}{Extra parameters.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{session}{Shiny session.}

\item{parent}{Parent global reactiveValues.}
}
\description{
Dashboard UI

Dashboard grid server
}
