% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_ui}
\alias{grid_ui}
\alias{grid_server}
\title{Dashboard grid UI}
\usage{
grid_ui(id)

grid_server(id, blocks, selected, mode, blocks_ns)
}
\arguments{
\item{id}{Module id.}

\item{blocks}{List of blocks.}

\item{selected}{Selected block.}

\item{mode}{App mode.}

\item{blocks_ns}{In which namespace are the blocks}
}
\description{
Dashboard grid UI

Dashboard grid server
}
