% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R, R/dock.R
\name{dashboard_ui}
\alias{dashboard_ui}
\alias{dashboard_server}
\alias{dashboard_ui.dag_board}
\alias{dashboard_server.dag_board}
\title{Dashboard Ui generic}
\usage{
dashboard_ui(id, board, ...)

dashboard_server(board, update, session, parent, ...)

\method{dashboard_ui}{dag_board}(id, board, ...)

\method{dashboard_server}{dag_board}(board, update, session, parent, ...)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board reactiveValues. Read-only.}

\item{...}{Extra parameters.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{session}{Shiny session.}

\item{parent}{Parent global reactiveValues.}
}
\description{
Dashboard Ui generic

Dashboard grid server

Dashboard dock UI method

Dashboard dock server method
}
