% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_ui}
\alias{grid_ui}
\alias{grid_server}
\title{Dashboard grid UI}
\usage{
grid_ui(id)

grid_server(id, parent, blocks, blocks_ns)
}
\arguments{
\item{id}{Module id.}

\item{parent}{Parent reactive values containing app mode
or which block should be in grid.}

\item{blocks}{Board blocks.}

\item{blocks_ns}{In which namespace are the blocks}
}
\description{
Dashboard grid UI

Dashboard grid server
}
