% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_ui}
\alias{grid_ui}
\alias{grid_server}
\title{Dashboard grid UI}
\usage{
grid_ui(id)

grid_server(id, board, mode, blocks_ns)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board object. Contains info about blocks,
selected block, refreshed status ...}

\item{mode}{App mode.}

\item{blocks_ns}{In which namespace are the blocks}
}
\description{
Dashboard grid UI

Dashboard grid server
}
