% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R, R/dock.R, R/grid.R
\name{dashboard_ui}
\alias{dashboard_ui}
\alias{dashboard_server}
\alias{dashboard_ui.dock_board}
\alias{dashboard_server.dock_board}
\alias{dashboard_ui.grid_board}
\alias{dashboard_server.grid_board}
\title{Dashboard Ui generic}
\usage{
dashboard_ui(id, board, ...)

dashboard_server(board, update, parent, ...)

\method{dashboard_ui}{dock_board}(id, board, ...)

\method{dashboard_server}{dock_board}(board, update, parent, ...)

\method{dashboard_ui}{grid_board}(id, board, ...)

\method{dashboard_server}{grid_board}(board, update, parent, ...)
}
\arguments{
\item{id}{Module id.}

\item{board}{Board reactiveValues. Read-only.}

\item{...}{Extra parameters.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{parent}{Parent global reactiveValues.}
}
\description{
Dashboard Ui generic

Dashboard grid server

Dashboard dock UI method

Dashboard dock server method

Dashboard grid UI method

Dashboard grid server method
}
