% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-code.R
\name{gen_code_ui}
\alias{gen_code_ui}
\title{Code generation module}
\usage{
gen_code_ui(id, board)
}
\arguments{
\item{board}{The initial \code{board} object}
}
\description{
Generate reproducible code from a board.
}
