% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{board_restore}
\alias{board_restore}
\alias{build_layout}
\alias{manage_scoutbar}
\title{Board restoration callback}
\usage{
board_restore(board, update, session, parent, ...)

build_layout(modules)

manage_scoutbar(board, update, session, parent, ...)
}
\description{
Board restoration callback

App layout

Scoutbar management callback
}
\keyword{internal}
