% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{manage_sidebars}
\alias{manage_sidebars}
\alias{manage_block_visibility}
\title{Manage board sidebars}
\usage{
manage_sidebars(rv, selected, session)

manage_block_visibility(parent, rv)
}
\arguments{
\item{rv}{Reactive values.}

\item{selected}{Selected node.}

\item{session}{Shiny session object.}
}
\description{
Manage board sidebars

Manage blocks visibility
}
\keyword{internal}
