% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{manage_block_visibility}
\alias{manage_block_visibility}
\alias{board_restore}
\alias{build_layout}
\alias{manage_scoutbar}
\title{Manage blocks visibility}
\usage{
manage_block_visibility(board, update, parent, ...)

board_restore(board, update, parent, ...)

build_layout(modules, session)

manage_scoutbar(board, update, parent, ...)
}
\description{
Manage blocks visibility

Board restoration callback

App layout

Scoutbar management callback
}
\keyword{internal}
