% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{manage_sidebars}
\alias{manage_sidebars}
\alias{manage_block_visibility}
\alias{manage_app_mode}
\alias{board_restore}
\title{Manage board sidebars}
\usage{
manage_sidebars(board, update, parent, ...)

manage_block_visibility(board, update, parent, ...)

manage_app_mode(board, update, parent, ...)

board_restore(board, update, parent, ...)
}
\arguments{
\item{board}{Board reactiveValues. Read-only.}

\item{update}{Update reactiveVal to signal change to the board.}

\item{parent}{Parent global reactiveValues.}

\item{...}{Extra elements.}
}
\description{
Manage board sidebars

Manage blocks visibility

Manage app mode

Board restoration callback
}
\keyword{internal}
