% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-board.R
\name{init_block_ui}
\alias{init_block_ui}
\alias{init_block_callbacks}
\title{Init block UI elements}
\usage{
init_block_ui(blk, session)

init_block_callbacks(blk, rv, input)
}
\arguments{
\item{input}{Shiny input object.}
}
\description{
Insert block UI in the DOM.

This initializes some observers needed
during the block lifecycle.
}
