% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{list_empty_connections}
\alias{list_empty_connections}
\title{List node connections}
\usage{
list_empty_connections(x, vals)
}
\arguments{
\item{x}{block.}

\item{vals}{Reactive values.}
}
\description{
List node connections
}
