% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/utils-board.R
\name{main_ui}
\alias{main_ui}
\alias{main_server}
\alias{grid_ui_wrapper}
\alias{properties_ui}
\alias{actions_ui}
\title{Main ui for blockr2}
\usage{
main_ui(id, board)

main_server(id, board)

grid_ui_wrapper(ns, width = "75\%", title = "Dashboard")

properties_ui(..., ns, width = "40\%", title = "Block properties")

actions_ui(..., ns)
}
\arguments{
\item{id}{Unique id.}

\item{board}{Board object.}

\item{ns}{Namespace}

\item{width}{Sidebar width.}

\item{title}{Sidebar title}

\item{...}{Extra UI elements.}
}
\description{
The board is composed of 2 views pointing to the network module
or the grid/dashboard module.

Server module for board.

Default board grid.

Default block sidebar.

Default action bar.
}
