% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{grid_ui}
\alias{grid_ui}
\alias{properties_ui}
\alias{actions_ui}
\alias{main_ui}
\alias{main_server}
\title{Board blocks grid}
\usage{
grid_ui(ns, width = "75\%", title = "Dashboard")

properties_ui(..., ns, width = "40\%", title = "Block properties")

actions_ui(..., ns)

main_ui(id)

main_server(id)
}
\arguments{
\item{ns}{Namespace}

\item{width}{Sidebar width.}

\item{title}{Sidebar title}

\item{...}{Extra UI elements.}

\item{id}{Unique id.}
}
\description{
Default board grid.

Default block sidebar.

Default action bar.

The board is composed of 2 views pointing to the network module
or the grid/dashboard module.

Server module for board.
}
