% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main_ui}
\alias{main_ui}
\alias{main_server}
\title{Main ui for blockr2}
\usage{
main_ui(id, board)

main_server(id, board)
}
\arguments{
\item{id}{Unique id.}

\item{board}{Board object.}
}
\description{
The board is composed of 2 views pointing to the network module
or the grid/dashboard module.

Server module for board.
}
