% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{register_node_validation}
\alias{register_node_validation}
\alias{apply_validation}
\title{Register block validation observer}
\usage{
register_node_validation(id, rv, vals, session)

apply_validation(message, id, vals, session)
}
\arguments{
\item{id}{Block id for which to register the validation.}

\item{rv}{Board reactive values. Read-only.}

\item{vals}{Global reactive values. Read-write.}

\item{session}{Shiny session object.}

\item{message}{Message.}
}
\description{
For each block we register an observer that
captures only the messages related to this block validation
status.

Block validation is made by the backend
this function only updates the node color
based on the valid status.
}
\keyword{internal}
