% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{custom_board_plugins}
\alias{custom_board_plugins}
\title{Register custom board plugins}
\usage{
custom_board_plugins(which = NULL)
}
\arguments{
\item{which}{(Optional) character vectors of plugins to include}
}
\description{
Ovewrite the default core plugins
}
