% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{process_grid_content}
\alias{process_grid_content}
\title{Process grid layout}
\usage{
process_grid_content(vals, grid_layout)
}
\arguments{
\item{vals}{Local module reactive values.}

\item{grid_layout}{Returned by input$<GRID_ID>_layout.
Contains blocks coordinates, dimensions, ...}
}
\description{
Process grid layout
}
\keyword{internal}
