% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{register_node_menu_obs}
\alias{register_node_menu_obs}
\title{Register observers related to the node menu}
\usage{
register_node_menu_obs(blocks_ids, vals, parent, obs, session)
}
\arguments{
\item{blocks_ids}{Board block ids.}

\item{vals}{Local scope reactive values.}

\item{parent}{Global scope reactive values.}

\item{obs}{Plugin observers list.}

\item{session}{Shiny session object.}
}
\description{
Observer to maintain the state between the 2 switches +
an observer to handle serialisation/restoration, to remove
nodes and to append to a given node.
}
\keyword{internal}
