% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-network.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove an edge}
\usage{
remove_edge(selected, rv)
}
\arguments{
\item{selected}{UID (character string) of edge to remove.}

\item{rv}{Reactive values containing dataframe representing edges data.}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
