% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove an edge}
\usage{
remove_edge(selected, vals, session)
}
\arguments{
\item{selected}{UID (character string) of edge to remove.}

\item{vals}{Reactive values containing dataframe representing edges data.}

\item{session}{Shiny session object.}
}
\description{
Update dataframe for visNetwork graph
}
\keyword{internal}
