% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{remove_node}
\alias{remove_node}
\title{Remove a node}
\usage{
remove_node(selected, vals, session)
}
\arguments{
\item{selected}{UID (character string) of node to remove.}

\item{vals}{Global reactive values.}

\item{session}{Shiny session object.}
}
\description{
Remove node from g6 instance
}
\keyword{internal}
