% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{restore_dashboard}
\alias{restore_dashboard}
\alias{find_blocks_ids}
\title{Restore dashboard state from board state}
\usage{
restore_dashboard(board, rv, parent, session)

find_blocks_ids(board, parent, session)
}
\arguments{
\item{board}{Board object.}

\item{rv}{Board reactive values object. Read-only}

\item{parent}{Parent reactive values.}

\item{session}{Shiny session object.
Contains blocks coordinates, dimensions, ...}
}
\description{
Restore dashboard state from board state

Find blocks ids
}
