% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{restore_dashboard}
\alias{restore_dashboard}
\alias{restore_dashboard.dash_board}
\alias{find_blocks_ids}
\alias{find_blocks_ids.dock_board}
\title{Restore dashboard state from board state}
\usage{
restore_dashboard(board, blocks, vals, parent, session)

\method{restore_dashboard}{dash_board}(board, blocks, vals, parent, session)

find_blocks_ids(board, parent, session)

\method{find_blocks_ids}{dock_board}(board, parent, session)
}
\arguments{
\item{board}{Board object.}

\item{blocks}{Board block objects.}

\item{vals}{Local reactive values.}

\item{parent}{Parent reactive values.}

\item{session}{Shiny session object.
Contains blocks coordinates, dimensions, ...}
}
\description{
Restore dashboard state from board state

Find blocks ids generic

Find blocks ids dock method
}
