% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dashboard.R
\name{restore_dashboard}
\alias{restore_dashboard}
\alias{restore_dashboard.dock_board}
\alias{restore_dashboard.grid_board}
\title{Restore dashboard state from board state}
\usage{
restore_dashboard(board, blocks, vals, parent, session)

\method{restore_dashboard}{dock_board}(board, blocks, vals, parent, session)

\method{restore_dashboard}{grid_board}(board, blocks, vals, parent, session)
}
\arguments{
\item{blocks}{Board block objects.}

\item{vals}{Local reactive values.}

\item{parent}{Parent reactive values.}

\item{session}{Shiny session object.
Contains blocks coordinates, dimensions, ...}
}
\description{
Restore dashboard state from board state

Restore grid state from board state

Restore grid state from board state
}
