% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{restore_grid}
\alias{restore_grid}
\title{Restore grid state from board state}
\usage{
restore_grid(board, blocks_ns, vals)
}
\arguments{
\item{board}{Board containing saved state.}

\item{blocks_ns}{Blocks namespace.}

\item{vals}{Local reactive values.
Contains blocks coordinates, dimensions, ...}
}
\description{
Restore grid state from board state
}
\keyword{internal}
