% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{restore_network}
\alias{restore_network}
\title{Restore network from saved snapshot}
\usage{
restore_network(links, vals, rv, session)
}
\arguments{
\item{links}{Board links.}

\item{vals}{Local reactive values.}

\item{rv}{Global parent reactive values.}

\item{session}{Shiny session object}
}
\value{
A reactiveValues object.
}
\description{
Network is updated via a proxy.
}
\keyword{internal}
