% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R, R/dag-board.R
\name{run_demo_app}
\alias{run_demo_app}
\alias{new_dag_board}
\alias{is_dag_board}
\alias{dag_board_options}
\title{Demo app}
\usage{
run_demo_app(...)

new_dag_board(
  ...,
  modules = list(new_dashboard_module(), new_chat_module()),
  options = dag_board_options(),
  class = character()
)

is_dag_board(x)

dag_board_options()
}
\arguments{
\item{...}{Forwarded to \link[blockr.core]{new_board}.}

\item{modules}{Further modules to pass.}

\item{options}{Board options (see \code{\link[=dag_board_options]{dag_board_options()}}).}

\item{class}{Additional class(es).}

\item{x}{(Board) object}
}
\description{
Run demo app
}
