% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{run_demo_app}
\alias{run_demo_app}
\alias{new_dag_board}
\title{Demo app}
\usage{
run_demo_app(...)

new_dag_board(..., modules = new_dashboard_module())
}
\arguments{
\item{...}{Forwarded to \link[blockr.core]{new_board}.}
}
\description{
Run demo app
}
