% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-serialize.R
\name{ser_deser_server}
\alias{ser_deser_server}
\alias{ser_deser_ui}
\title{Serialization module}
\usage{
ser_deser_server(id, board, parent, ...)

ser_deser_ui(id, board)
}
\arguments{
\item{id}{module ID.}

\item{board}{The initial \code{board} object}

\item{parent}{App state}

\item{...}{Extra arguments passed from parent scope}
}
\value{
A \link[shiny]{reactiveVal} object that evaluates to \code{NULL} or a
\code{board} obejct.
}
\description{
Object (de)serialization in a board server context.
}
