% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-serialize.R
\name{ser_deser_ui}
\alias{ser_deser_ui}
\title{Ser/deser module UI}
\usage{
ser_deser_ui(id, board)
}
\arguments{
\item{id}{module ID.}

\item{board}{The initial \code{board} object}
}
\description{
Ser/deser module UI
}
