% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ai-tools.R
\name{setup_chat_provider}
\alias{setup_chat_provider}
\title{Utility functions to create AI tools for blockr.ui}
\usage{
setup_chat_provider(
  chat = blockr_option("chat_function", default_chat),
  prompt = readLines(system.file("examples/ai-chat/rules.md", package = "blockr.ui")),
  ...
)
}
\arguments{
\item{chat}{\code{\link[ellmer:chat-any]{ellmer::chat()}}-like function.}

\item{prompt}{System prompt to use for the AI provider. Best
to create a markdown file with instructions.}

\item{...}{Additional arguments to pass to the provider.}
}
\description{
Utility functions to create AI tools for blockr.ui
}
\keyword{internal}
