% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-links.R
\name{show_stack_actions}
\alias{show_stack_actions}
\title{Show stack actions}
\usage{
show_stack_actions(rv, session)
}
\arguments{
\item{rv}{Reactive values containing board elements. Read-only.}

\item{session}{Shiny session object.}
}
\description{
A modal window triggered when create stack is
pressed on canvas right click
}
\keyword{internal}
